==> 03 Scripts/01 Data Wrangling & Subsetting.ipynb <==
## 4.4 Data Wrangling & Subsetting

### 4.4.1 Import libraries & datasets

import pandas as pd
import numpy as np
import os

project_path = r'C:\Users\cneva\06-2024 Instacart Basket Analysis'
df_prods = pd.read_csv(os.path.join(project_path,'02 Data','Original Data', 'products.csv'))
df_ords = pd.read_csv(os.path.join(project_path,'02 Data','Original Data', 'orders.csv'))
df_dep = pd.read_csv(os.path.join(project_path, '02 Data','Original Data','departments.csv'))

### 4.4.2 Dropping columns

df_ords = df_ords.drop(columns = ['eval_set'])

df_ords['days_since_prior_order'].value_counts(dropna = False)

### 4.4.3 Renaming Columns

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

### 4.4.3 Changing a Variable's Data Type

df_ords['order_id'] = df_ords['order_id'].astype('str')

### 4.4.4 Transposing Data

df_dep_t = df_dep.T

df_dep_t.reset_index()

new_header = df_dep_t.iloc[0]

df_dep_t_new = df_dep_t[1:]

df_dep_t_new.columns = new_header

### 4.4.5 Data Dictionaries

data_dict = df_dep_t_new.to_dict('index')

### 4.4.6 Subsetting

df_snacks = df_prods[df_prods['department_id'] == 19]
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

### 4.4.7 Exporting Dataframes

# df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

# 4.4 Task

#### 2. Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.

df_ords['user_id'] = df_ords['user_id'].astype('str')

#### 3. Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the dataframe.

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = False)

#### 4. Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.
#### - 10 am is the busiest hour of the day

df_ords['order_hour_of_day'].value_counts(dropna = False)

#### 6. The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.

df_dep_t_new[df_dep_t_new['department'] == 'breakfast']

df_prods_breakfast = df_prods[df_prods['department_id'] == 14]

#### 7. They’d also like to see details about products that customers might use to throw dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.

df_dep_t_new[df_dep_t_new['department'].isin(['alcohol','deli','beverages','meat seafood'])]

df_prods_dinner_parties = df_prods[df_prods['department_id'].isin([5,7,12,20])]

#### # 8. It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?
#### - There are 7650 rows

df_prods_dinner_parties.shape
# There are 7650 rows

#### 9. Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.”  Extract all the information you can about this user.
#### - user_id 1 always places orders at the beginning of the week, usually early in the morning / early afternoon.  Orders are placed every 2-4 weeks apart

df_ords_customer1 = df_ords[df_ords['user_id'] == '1']
df_ords_customer1

df_ords_customer1.describe()

#### 12. Export your df_ords dataframe as “orders_wrangled.csv” in your “Prepared Data” folder.

df_ords.to_csv(os.path.join(project_path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(project_path,'02 Data', 'Prepared Data', 'departments_wrangled.csv'))
