==> 03 Scripts/02 Data Cleaning.ipynb <==
## 

import pandas as pd
import numpy as np
import os

project_path = r'C:\Users\cneva\06-2024 Instacart Basket Analysis'
df_prods = pd.read_csv(os.path.join(project_path,'02 Data','Original Data', 'products.csv'))
df_ords = pd.read_csv(os.path.join(project_path,'02 Data','Prepared Data', 'orders_wrangled.csv'), index_col = 0)
df_dep = pd.read_csv(os.path.join(project_path, '02 Data','Prepared Data','departments_wrangled.csv'))

#### 4.5.1 Mixed-Type Data 

# Create a dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test

# Check for mixed tyeps
for col in df_test.columns.tolist():
    weird = (df_test[[col]].map(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
    if len(df_test[weird]) > 0:
        print(col)

df_test['mix'] = df_test['mix'].astype('str')

df_test['mix'].dtype

#### 4.5.2 Missing values

df_prods.isnull().sum()

df_nan = df_prods[df_prods['product_name'].isnull() == True]
df_nan

# Replace missing values with the mean:
# df['column with missings'].fillna(mean value, inplace=True)

# replace missing values with the median:
# df['column with missings'].fillna(median value, inplace=True)

# create new df with non missing rows
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# Drop all missing values
# df_prods.dropna(inplace = True)

# Only drop missing valuse from a particular column
# df_prods.dropna(subset = [‘product_name’], inplace = True)
# inplace = True = overwrites original dataframe
# inplace = False = return a view of the changed dataframe

#### 4.5.3 Duplicates 

# Look for full duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

df_prods_clean.shape

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups.shape

# df_prods_clean_no_dups.to_csv(os.path.join(project_path, '02 Data', 'Prepared Data', 'products_clean.csv'))

### 4.5 Task 

##### 2. Run the df.describe() function on your df_ords dataframe. Using your new knowledge about how to interpret the output of this function, share in a markdown cell whether anything about the data looks off or should be investigated further.

##### - order_number has a max number of 100 which may indicate that the data is for custoemrs with <= 100 orders
##### - order_dow has a min of 0 and max of 6 which represents the 7 days of the week
##### - order_hour_of_day has a min of 0 and max of 23 which represents the 24 hours of hte day
##### - days_since_prior_order has a max number of 30 which may indicate that the data is for customers who consistently place orders within 30 days

df_ords.describe()

##### 3. Check for mixed-type data in your df_ords dataframe. 

# for col in df_ords.columns.tolist():
#     weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
#     if len(df_ords[weird]) > 0:
#         print(col)

# Check if data framme has mixed data types within the same column. False = none
mixed_types_per_column = df_ords.apply(lambda x: len(x.apply(type).unique()) > 1)
print(mixed_types_per_column)

##### 4. If you find mixed-type data, fix it. The column in question should contain observations of a single data type.
##### - There are no mixed-type columns

##### 5. Run a check for missing values in your df_ords dataframe.

df_ords.isnull().sum()

##### 6. Address the missing values using an appropriate method.
##### - It looks like an empty value in the days_since_prior_order means that it is that user's first and only order since the number of observations where days_since_prior_order is missing is the same as the number of unique customers for that subset of data.

df_ords[df_ords['days_since_prior_order'].isna()]

len(df_ords['user_id'].unique())

##### Another check to confirm is to see if all order_numbers are equal to 1

df_ords_empty = df_ords[df_ords['days_since_prior_order'].isna()]
df_ords_empty[df_ords_empty['order_number'] != 1]

##### Nothing should be done to the days_since_prior_order column as deleting / imputing values would altar the overal dataset drastically.  It would be prudent to add another column that might explain the emptys such as a first order flag.  This would alleviate any confusion if any other person were to do further consistency checks on this dataframe.

df_ords['first_order'] = df_ords['days_since_prior_order'].isna()
df_ords.head()

##### 7.  Run a check for duplicate values in your df_ords data.
##### - There are no duplicated values

df_ords[df_ords.duplicated()]

##### 8. Address the duplicates using an appropriate method.
##### - No full duplicates

##### 9. Export your final, cleaned df_prods and df_ords data as “.csv” files in your “Prepared Data” folder and give them appropriate, succinct names.

df_prods_clean_no_dups.to_csv(os.path.join(project_path, '02 Data', 'Prepared Data', 'products_clean.csv'))
df_ords.to_csv(os.path.join(project_path, '02 Data', 'Prepared Data', 'orders_clean.csv'))
