==> 03 Scripts/03 Join Customers, Orders,.ipynb <==
###  4.6 Combining & Exporting Data

import pandas as pd
import numpy as np
import os

project_path = r'C:\Users\cneva\06-2024 Instacart Basket Analysis'


df_prod = pd.read_csv(os.path.join(project_path, '02 Data', 'Prepared Data', 'products_clean.csv'), index_col = 0)
df_ords = pd.read_csv(os.path.join(project_path, '02 Data', 'Prepared Data', 'orders_clean.csv'), index_col = 0)
df_dep = pd.read_csv(os.path.join(project_path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'), index_col = 0)
df_ords_prior = pd.read_csv(os.path.join(project_path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = 0)

df_ords_prior.head()

df_ords_prior.shape

df_ords_prior['reordered'].value_counts()

df_ords_prior.describe()

### 4.6 Task

#### 1. Merge df_ords & df_ords_prior 

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

#### 2.  Export the merged file in pickle format as “orders_products_combined.pkl”.

df_merged_large.to_pickle(os.path.join(project_path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))
