==> 03 Scripts/04 Join Customers-Order, Products.ipynb <==
import pandas as pd
import numpy as np
import os

project_path = r'C:\Users\cneva\06-2024 Instacart Basket Analysis'

### 4.6 Task

#### 3. In a new notebook, import the orders_products_combined dataframe from the pickle file you just saved.

df_orders_products = pd.read_pickle(os.path.join(project_path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_prods = pd.read_csv(os.path.join(project_path, '02 Data', 'Prepared Data', 'products_clean.csv'), index_col = 0)

#### 4. Check the shape of the imported dataframe (it should be the same as the one you exported—always check!).

df_orders_products.shape

#### 5. Determine a suitable way to combine the orders_products_combined dataframe with your products data set. Make sure you’re using your wrangled, cleaned, and deduped products data set stored in your “Prepared Data” folder from the previous Exercise’s task.

df_prods.head()

df_orders_products.head()

df_orders_products = df_orders_products.drop(columns = ['_merge'])

df_ords_prods_merged = df_orders_products.merge(df_prods, on = 'product_id', indicator = True, how = 'outer')

#### 6. Confirm the results of the merge using the merge flag.

df_ords_prods_merged['_merge'].value_counts()

 df_ords_prods_merged[df_ords_prods_merged['_merge'].isin(['left_only'])].head(3)

 df_ords_prods_merged[df_ords_prods_merged['_merge'].isin(['right_only'])].head(3)

##### There are 30,200 rows that are missing in the prods tables.  These are probably the products that were missing values and were dropped when cleaning the products table.
##### There are 11 rows that are missing from the ords_prods_prior_merge table.  These are probably products that have never been purchased
##### Because of the relatively low mismatches and we know the source/reason for the missing data an inner join will suffice.

df_ords_prods_merged = df_orders_products.merge(df_prods, on = 'product_id', indicator = True, how = 'inner')

df_ords_prods_merged.shape

df_ords_prods_merged.head(3)

df_ords_prods_merged['_merge'].value_counts(dropna = False)

df_ords_prods_merged = df_ords_prods_merged.drop(columns = ['_merge'])

#### 7. Export the newly created dataframe as ords_prods_merge in a suitable format (taking into consideration the size).

df_ords_prods_merged.to_pickle(os.path.join(project_path, '02 Data', 'Prepared Data', 'orders_and_products_merged_2.pkl'))

#### 8. Ensure your notebooks and Instacart project folder are organized and that comments and section headings have been used throughout your code. All your exported data files should be effectively labeled and stored in your “Data” folder.

#### 9. Save the two notebooks and send them to your tutor along with a screenshot of the exported data sets in your Instacart project folder.
