==> 03 Scripts/05 Feature Engineering.ipynb <==
### Exercise 4.7 Deriving New Variables

import pandas as pd
import numpy as np
import os

project_path = r'C:\Users\cneva\06-2024 Instacart Basket Analysis'
df_ords_prods = pd.read_pickle(os.path.join(project_path,'02 Data', 'Prepared Data', 'orders_and_products_merged_2.pkl'))

df_ords_prods.shape

df_ords_prods_short = df_ords_prods[:1000000]

df_ords_prods_short.shape

#### If-Statements with User-Defined Functions

# Define a function using if statements that categorizes products by the price 
def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range product'
    elif row['prices'] > 15:
        return 'High-range product'
    else: return 'Not enough data'

# Apply the price_range function on newly created column price_range
# axis = 1 = apply function on all rows, axis = 0 = apply function on all columns
df_ords_prods_short['price_range'] = df_ords_prods_short.apply(price_label, axis = 1)

df_ords_prods_short.head()

df_ords_prods_short['price_range'].value_counts(dropna = False)

#### If-Statements with the loc() Function 

df_ords_prods_short.loc[df_ords_prods['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_short.loc[(df_ords_prods['prices'] <= 15) & (df_ords_prods['prices'] > 5) , 'price_range_loc'] = 'Mid-range product'

df_ords_prods_short.loc[df_ords_prods['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_short['price_range_loc'].value_counts(dropna = False)

##### If-Statements with For-Loops

# create empty list to store results of the for look
result = []

# loop through each value/row in 'prices' and store result in list
for x in df_ords_prods_short['prices']:
    if x <= 5:
        result.append('Low-range product')
    elif x <= 15 and x > 5:
        result.append('Mid-range product')
    elif x > 15:
        result.append('High-range product')
    else:
        'Not enough data'

# create new column using the result list
df_ords_prods_short['price_range_for'] = result

#### 4.7 Task

##### 1.  If you haven’t done so already, complete the instructions in the Exercise for creating the “price_label” and “busiest_day” columns.

# create price categories using .loc
df_ords_prods.loc[df_ords_prods['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods.loc[(df_ords_prods['prices'] <= 15) & (df_ords_prods['prices'] > 5) , 'price_range_loc'] = 'Mid-range product'

df_ords_prods.loc[df_ords_prods['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods['price_range_loc'].value_counts(dropna = False)

# create busiest_day categories using for loop
df_ords_prods['orders_day_of_week'].value_counts(dropna = False)

# create list to store results
busy_result = []

# loop through orders_day_of_week and assign busy category
for x in df_ords_prods['orders_day_of_week']:
    if x == 0:
        busy_result.append('Busiest day')
    elif x == 4:
        busy_result.append('Least busy')
    else:
        busy_result.append('Regularly busy')

# create busy column using list
df_ords_prods['busiest_day'] = busy_result

df_ords_prods['busiest_day'].value_counts(dropna = False)

##### 2.  Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. Now, they want “Busiest day” to become “Busiest days” (plural). This label should correspond with the two busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know the two slowest days. Create a new column for this using a suitable method.

# create list to store results
busy_result_2 = []

# loop through orders_day_of_week and assign busy category
for x in df_ords_prods['orders_day_of_week']:
    if x == 0 or x == 1:
        busy_result_2.append('Busiest days')
    elif x == 4 or x == 3:
        busy_result_2.append('Least busy days')
    else:
        busy_result_2.append('Regularly busy days')

# create busy column using list
df_ords_prods['busiest_day_2'] = busy_result_2

##### 3. Check the values of this new column for accuracy. Note any observations in markdown format.


df_ords_prods['busiest_day_2'].value_counts(dropna = False)

#####   - The distribution between the buckets are more even because the number of days within each category has been spread; 1 day, 1 day, 5 days vs 2 days, 2 days, 3 days. We expect the Regularly busy days to be the most because it's 3 days.  And we expect the Busiest days to be more than least busy days because obviously there would be more orders therefore more observations.

##### 4. When too many users make Instacart orders at the same time, the app freezes. The senior technical officer at Instacart wants you to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” Create a new column containing these labels called “busiest_period_of_day.”

# identify the hours when most  busy
df_ords_prods['order_hour_of_day'].value_counts(dropna = False)

###### 8-hour buckets: 
###### Most orders -    8 - 16
###### Average orders -  17 - 23
###### Fewest orders -  0 - 7

# check if buckets make sense
df_ords_prods[(df_ords_prods['order_hour_of_day'] >= 0) & (df_ords_prods['order_hour_of_day'] < 8)].shape

df_ords_prods[(df_ords_prods['order_hour_of_day'] >= 8) & (df_ords_prods['order_hour_of_day'] < 17)].shape

df_ords_prods[(df_ords_prods['order_hour_of_day'] >= 17) & (df_ords_prods['order_hour_of_day'] < 24)].shape

# Most orders = order_hour_of_day >= 8, < 17
df_ords_prods.loc[(df_ords_prods['order_hour_of_day'] >= 8) & (df_ords_prods['order_hour_of_day'] < 17) , 'busiest_period_of_day'] = 'Most orders'

# Average orders = order_hour_of_day >= 17, < 24
df_ords_prods.loc[(df_ords_prods['order_hour_of_day'] >= 17) & (df_ords_prods['order_hour_of_day'] < 24) , 'busiest_period_of_day'] = 'Average orders'

# Fewest orders = order_hour_of_day >= 0, < 7
df_ords_prods.loc[(df_ords_prods['order_hour_of_day'] >= 0) & (df_ords_prods['order_hour_of_day'] < 8) , 'busiest_period_of_day'] = 'Fewest orders'

##### 5. Print the frequency for this new column.

df_ords_prods['busiest_period_of_day'].value_counts(dropna = False)

df_ords_prods.head(1)

##### 6.  Ensure your notebook is clean and structured and that your code is well commented.

df_ords_prods.to_pickle(os.path.join(project_path, '02 Data', 'Prepared Data', 'ords_prods_3.pkl'))

##### 7.  Export your dataframe as a pickle file (since you added new columns) and store it correctly in your “Prepared Data” folder.

##### 8.  Save your notebook and submit it to your tutor for review.
