==> 03 Scripts/06 Grouping & Aggregating Variables.ipynb <==
### 4.8 Task 

#### 1. Create a new notebook for this task. Be sure to import the relevant libraries, along with your ords_prods_merge dataframe, which should include your newly derived columns from the previous Exercise.

import pandas as pd
import numpy as np
import os

project_path = r'C:\Users\cneva\06-2024 Instacart Basket Analysis'
df_ords_prods = pd.read_pickle(os.path.join(project_path, '02 Data', 'Prepared Data', 'ords_prods_3.pkl'))

df_ords_prods.shape

df_ords_prods.head(1)

#### 2. In this Exercise, you learned how to find the aggregated mean of the “order_number” column grouped by “department_id” for a subset of your dataframe. Now, repeat this process for the entire dataframe.

df_ords_prods.groupby('department_id').agg({'order_number': ['mean']})

#### 3. Analyze the result. How do the results for the entire dataframe differ from those of the subset? Include your comments in a markdown cell below the executed code.

##### The subset didnt include all departments. Because the dataframe contains much more data, some departments have gone up and some have gone down

#### 4. Follow the instructions in the Exercise for creating a loyalty flag for existing customers using the transform() and loc() functions.

# crate max order per user_id column
df_ords_prods['max_order'] = df_ords_prods.groupby(['user_id'])['order_number'].transform(np.max)

df_ords_prods.head(2)

# create loyalty category column: 
# loyal customer = max order > 40
# regular customer = 10 < max order <= 40
# new customer = max order <= 10
df_ords_prods.loc[df_ords_prods['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

df_ords_prods.loc[(df_ords_prods['max_order'] > 10) & (df_ords_prods['max_order'] <= 40), 'loyalty_flag'] = 'Regular customer'

df_ords_prods.loc[df_ords_prods['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

df_ords_prods.head(5)

df_ords_prods['loyalty_flag'].value_counts(dropna = False)

#### 5. The marketing team at Instacart wants to know whether there’s a difference between the spending habits of the three types of customers you identified. Use the loyalty flag you created and check the basic statistics of the product prices for each loyalty category (Loyal Customer, Regular Customer, and New Customer). What you’re trying to determine is whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers.

# groupby loyalty flag and aggregate mean, min and max
df_ords_prods.groupby('loyalty_flag').agg({'prices':['mean','min','max']})

##### On average, new customers purchase higher priced products while loyal customer purchase the lower priced products

#### 6. The team now wants to target different types of spenders in their marketing campaigns. This can be achieved by looking at the prices of the items people are buying. Create a spending flag for each user based on the average price across all their orders using the following criteria:
##### - If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender.”
##### - If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a “High spender.”

# Groupby user_id and aggregate by average price
df_ords_prods['user_average_price'] = df_ords_prods.groupby('user_id')['prices'].transform(np.mean)

# Check 2 users
df_ords_prods[df_ords_prods['user_id'].isin([99,9999])]

# flag user based on user_average_price
# lower spender = user_average_price < 10
# high spender = user_avereage_price >= 10
df_ords_prods.loc[df_ords_prods['user_average_price'] < 10, 'user_spender_type'] = 'Low spender'

df_ords_prods.loc[df_ords_prods['user_average_price'] >= 10, 'user_spender_type'] = 'High spender'

# check flag
df_ords_prods.head(3)

df_ords_prods['user_spender_type'].value_counts(dropna = False)

#### 7. In order to send relevant notifications to users within the app (for instance, asking users if they want to buy the same item again), the Instacart team wants you to determine frequent versus non-frequent customers. Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. The criteria for the flag should be as follows:
##### - If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”
##### - If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
##### - If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”

# groupby user_id and aggregate days_since_prior_order by the median
df_ords_prods['user_median_days_since'] = df_ords_prods.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# flag user based on user_average_price
# Non-frequent customer = user_median_days_since > 20
# Regular customer = 10 < user_median_days_since <= 20
#  Frequent customer = user_median_days_since <= 10
df_ords_prods.loc[df_ords_prods['user_median_days_since'] > 20, 'user_order_frequency_flag'] = 'Non-frequent customer'

df_ords_prods.loc[(df_ords_prods['user_median_days_since'] <= 20) & (df_ords_prods['user_median_days_since'] > 10), 'user_order_frequency_flag'] = 'Regular customer'

df_ords_prods.loc[df_ords_prods['user_median_days_since'] <= 10, 'user_order_frequency_flag'] = 'Frequent customer'

df_ords_prods.head(3)

df_ords_prods['user_order_frequency_flag'].value_counts()

#### 8. Ensure your notebook is clean and structured and that your code is well commented.

#### 9. Export your dataframe as a pickle file and store it correctly in your “Prepared Data” folder.

df_ords_prods.to_pickle(os.path.join(project_path,'02 Data','Prepared Data','ords_prods_4.pkl'))

#### 10. Save your notebook and submit it to your tutor for review.
