==> 03 Scripts/07 Data Visualization v1.ipynb <==
### Exercise 3.9 Task Part 1

#### 01 Import libraries and files

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = r'C:\Users\cneva\06-2024 Instacart Basket Analysis'

df_cust = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

#### 02 Check columns and shape

df_cust.shape

df_cust.head()

# drop first and last name as this is personal information and probably won't be needed for analysis
df_cust = df_cust.drop(columns = ['First Name', 'Surnam'])

# normalize column names for clarity
df_cust.rename(columns = {'Gender':'gender', 'STATE':'state', 'Age':'age'}, inplace = True)

#### 03 Check high level statistics for anything unusual

df_cust.describe()

#### 04 Check for improperly assigned data types 

df_cust.dtypes

df_cust['user_id'] = df_cust['user_id'].astype('str')

df_cust.dtypes

#### 05 Check for mixed data types

for col in df_cust.columns.tolist():
    if df_cust[col].apply(type).nunique() > 1:
        print(col)

#### 06 Look for full duplicates

df_cust[df_cust.duplicated()]

#### 07 Look for missing values

df_cust.isnull().sum()

#### 08 Import Instacart data 

df_ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_4.pkl'))

df_ords_prods.head()

df_ords_prods.dtypes

df_ords_prods['user_id'] = df_ords_prods['user_id'].astype('str')

# check if df_ords_prods have any user_ids that are not in df_cust
df_ords_prods[~df_ords_prods['user_id'].isin(df_cust['user_id'])]['user_id'].count()

# check if df_cust have any user_ids that are not in df_ords_prods
df_cust[~df_cust['user_id'].isin(df_ords_prods['user_id'])]['user_id'].count()

#### 09 Merge

# merge
df_instacart = df_ords_prods.merge(df_cust, on = 'user_id')

df_instacart.head()

df_instacart.shape

#### 10 Export 

df_instacart.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_cust.pkl'))
